chore(runner): refactor
