+++ runner/src/managers/mod.rs
@@ -1,4 +1,4 @@
 mod connection;
-mod tasks;
+pub(crate) mod tasks;
 
 pub(crate) use connection::ConnectionManager;
